[Kernel][Drivers] Make Keyboard a CharacterDeviceDriver
